refactor: Move the random edit case generation functions to their own file
